pyproject.toml
[project]
name = "cloudprojektvistula2026"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.129.0",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.41.0",
    "psycopg[binary]>=3.1.18",
    "jinja2>=3.1.6",
    "azure-cosmos>=4.15.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "ruff>=0.15.2",
]

[tool.ruff]
line-length = 120
target-version = "py314"

[tool.ruff.lint]
# E: pycodestyle errors
# W: pycodestyle warnings
# F: Pyflakes
# I: isort (sortowanie importów)
# B: flake8-bugbear (częste błędy)
select = ["E", "W", "F", "I", "B"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
